+++ package.json
@@ -5,18 +5,24 @@
     "dependencies": {
         "@material-ui/core": "^1.5.1",
         "@material-ui/icons": "^1.1.1",
+        "acl": "^0.4.11",
         "ajv": "^6.9.1",
         "axios": "^0.18.0",
+        "bcrypt": "^3.0.6",
         "body-parser": "^1.18.3",
         "chromedriver": "^2.46.0",
         "express": "^4.16.4",
         "express-pino-logger": "^4.0.0",
         "geckodriver": "^1.15.1",
+        "generate-password": "^1.4.2",
+        "jsonwebtoken": "^8.5.1",
+        "mongoose": "^5.6.4",
         "mysql": "^2.16.0",
         "node-env-run": "^3.0.2",
         "nodemailer": "^6.0.0",
         "nodemon": "^1.18.10",
         "npm-run-all": "^4.1.5",
+        "password-hash": "^1.2.2",
         "pino-colada": "^1.4.4",
         "prop-types": "^15.7.2",
         "react": "^16.8.2",
@@ -27,7 +33,9 @@
         "react-test-renderer": "^16.8.3",
         "selenium": "^2.20.0",
         "selenium-webdriver": "^4.0.0-alpha.1",
-        "uuid": "^3.3.2"
+        "uninstall": "0.0.0",
+        "uuid": "^3.3.2",
+        "validator": "^11.1.0"
     },
     "devDependencies": {
         "babel-eslint": "^10.0.1",
