+++ server/config/index.js
@@ -1,0 +1,9 @@
+module.exports =
+        {
+            db:
+                    {
+                        uri: 'mongodb://' + ('localhost:27017') + '/xroads'
+                    },
+            secret: 'ilovecoding',
+            apiPath:'/api'
+        }
