+++ server/config/startup.js
@@ -1,0 +1,17 @@
+module.exports =
+        {
+            roles:
+                    {
+                        0: 'Founder',
+                        1: 'Co-Founder',
+                        2: 'Memeber',
+                        3: 'other User'
+                    },
+            stages: {
+                0: 'Idea',
+                1: 'Market Research',
+                2: 'Production',
+                3: 'MVP',
+                4: 'Other'
+            }
+        }
