+++ server/routes/index.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/', function(req, res, next) {
+  res.send({data:true})
+});
+module.exports = router;
