+++ server/routes/user.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+const userController = require('../controllers/user.controller');
+
+
+router.post('/signup', userController.signUp);
+router.post('/signin', userController.signIn);
+
+module.exports = router;
